+++ it/content/src/main/content/jcr_root/content/forms/af/core-components-it/samples/datalayer/basic/.content.xml
@@ -5,7 +5,7 @@
         cq:deviceGroups="[mobile/groups/responsive]"
         cq:lastModified="{Date}2023-03-28T15:01:13.482+05:30"
         cq:lastModifiedBy="admin"
-        cq:template="/conf/core-components-it/settings/wcm/templates/af-blank-v2"
+        cq:template="/conf/core-components-examples/settings/wcm/templates/af-blank-v2"
         jcr:language="en"
         jcr:primaryType="cq:PageContent"
         jcr:title="Adaptive Form V2 (IT)"
